Update changelog for release 0.4.5 and adjust Gemfile dependencies
This commit updates the CHANGELOG to include details for the newly
released version 0.4.5. The release notes detail the addition of
features such as obtaining a list of available models from an
Ollama server and checking model existence via new methods.

Additionally, the Gemfile has been updated to include the required
`securerandom` gem and other dependency versions have been adjusted
in the Gemfile.lock for compatibility and performance improvements.